requirements.txt
opencv-python
mediapipe
pygame
ultralytics
numpy
pyyaml
pyserial
